requirements.txt
# 核心依赖（命令行工具 + NoneBot2 插件均需要）
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
prettytable>=3.0.0
chardet>=4.0.0
pydantic>=2.0.0

# NoneBot2 插件依赖
nonebot2>=2.3.0
nonebot-adapter-onebot>=2.4.0

# 命令行工具额外依赖（B50 图片生成、曲线分析等）
matplotlib>=3.4.0
scipy>=1.7.0
Pillow>=8.0.0
